try3

+++ palindrome_integer/0-is_palindrome.c
@@ -4,12 +4,13 @@
 #include "palindrome.h"
 
 
-int is_palindrome(unsigned long long number)
+int is_palindrome(unsigned long number)
 {
     if(number == ULONG_MAX)
     {
         return (0);
     }
+
     int c = 0; /* digit position */
     long long n; /* number */
     n = number;
